Change collect() to receive y as a double instead of int

+++ plugin-api/src/main/java/plugin_api/YReceiver.java
@@ -11,5 +11,5 @@ public interface YReceiver {
      * 
      * @param y The calculated y value
      */
-    void collect(int y);
+    void collect(double y);
 }
